+++ mod/data.lua
@@ -1,0 +1,8 @@
+data:extend{
+	{
+		type = "custom-input",
+		name = "debugadapter-restart-session",
+		key_sequence = "CONTROL + SHIFT + F5",
+		action = "lua",
+	}
+}